not compiling, trying to solve srchitect issues cleanly for the "system". The transfer from drawing_output_t to display_context_visual_t as a stand alone base class will resolve circular dependency that occurred before.
A more distinct and pipeline dialect combined with base class
inheritance of the functionality where necessary can ease new object
creation.

Pipeline archecture allows for definietion of one function which may be
used to draw the object. from this base, the generation of the context
functions clip, and cached fn,clip can be accaomplished.

The dryness of the platform design "display_unit" is provided for by
separating the units to individual file.