source back not compiling
The visitor pattern used in the pipeline execute was not compiling
correctly. The errors were exhaustive and the answer was not apparent.
However, understandable. The function prototypes for fn_emit_XXX must be
unique within the variant. While the variant does not complain, the
visitation had a compile issue.

Hence the abstract classes for relative coordinate now use the single
fn_emit_cr_a_t instead of having their own. This will have to be
accounted for in signature of the type index used to match the interface
and store it within the map. Otherwise multiple functions will occupy
the same space.