compiles - non functional
Current the system has four major areas that need work.

1) visitor_interfaces_t object is not storing the published interfaces.
This is because the template constructor is not being called rather the
default one is. An instaniation issue within how the display unit object
is described within the hpp file.

2) A sequence needs to be maintained in addtion to the order stage. That
is each order stage needs a sort order for how they are inserted.

3) ink areas of drawing objects is not acquiring the emit functions.
Need to decide how to invoke them properly when adding a visual to the
display context.

4) change the pipeline visitor attributes to use bit masks for the
linkages copied from the diaply context on rendering objects.

5) decide which flags to copy for specific settings of a rendering
object. For example textual render changes based upon the path layout or
normal render.

6) Perhaps the use of weak pointers may have to be used as objects are
disconnected from the system.

7) multithreaded testing.

8) hashing object changes.
etc....